fix: move activity card test helpers to feature-local test-support
Relocate current-activity-card harness/mocks from global src/testing into
workflow-activity test-support and skip those paths in the feature-boundary
guard so PRD-local builders stay colocated without Vitest discovery issues.
