+++ ui/tsconfig.app.json
@@ -27,9 +27,8 @@
     "src/**/*.test.ts",
     "src/**/*.test.tsx",
     "src/**/*.stories.tsx",
+    "src/features/**/test-support/**",
     "src/features/bento/components/dashboard-bento-story-shared.tsx",
-    "src/testing/react-flow-current-activity-card-component.harness.tsx",
-    "src/testing/react-flow-current-activity-card-component.mocks.ts",
     "packages/components/src/**/*.test.ts",
     "packages/components/src/**/*.test.tsx",
     "packages/components/src/**/*.harness.test.ts",
